Move keyring helpers into their own module
