+++ primitives/test-utils/src/lib.rs
@@ -19,14 +19,17 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 use bp_header_chain::justification::GrandpaJustification;
-use ed25519_dalek::{Keypair, PublicKey, SecretKey, Signature, Signer};
-use finality_grandpa::voter_set::VoterSet;
-use sp_application_crypto::{Public, TryFrom};
-use sp_finality_grandpa::{AuthorityId, AuthorityList, AuthorityWeight};
+use ed25519_dalek::Signer;
+use sp_application_crypto::TryFrom;
+use sp_finality_grandpa::{AuthorityId, AuthorityWeight};
 use sp_finality_grandpa::{AuthoritySignature, SetId};
 use sp_runtime::traits::{Header as HeaderT, One, Zero};
-use sp_runtime::RuntimeDebug;
 use sp_std::prelude::*;
+
+// Re-export all our test account utilities
+pub use keyring::*;
+
+mod keyring;
 
 pub const TEST_GRANDPA_ROUND: u64 = 1;
 pub const TEST_GRANDPA_SET_ID: SetId = 1;
@@ -218,85 +221,4 @@ pub fn test_header<H: HeaderT>(number: H::Number) -> H {
 /// Convenience function for generating a Header ID at a given block number.
 pub fn header_id<H: HeaderT>(index: u8) -> (H::Hash, H::Number) {
 	(test_header::<H>(index.into()).hash(), index.into())
-}
-
-pub trait Keyring {
-	fn public(&self) -> PublicKey {
-		(&self.secret()).into()
-	}
-
-	fn secret(&self) -> SecretKey;
-
-	fn pair(&self) -> Keypair {
-		let mut pair: [u8; 64] = [0; 64];
-
-		let secret = self.secret();
-		pair[..32].copy_from_slice(&secret.to_bytes());
-
-		let public = self.public();
-		pair[32..].copy_from_slice(&public.to_bytes());
-
-		Keypair::from_bytes(&pair).expect("We expect the SecretKey to be good, so this must also be good.")
-	}
-
-	fn sign(&self, msg: &[u8]) -> Signature {
-		self.pair().sign(msg)
-	}
-}
-
-#[derive(RuntimeDebug, Clone, Copy)]
-pub struct Account(pub u8);
-
-impl Keyring for Account {
-	fn secret(&self) -> SecretKey {
-		SecretKey::from_bytes(&[self.0; 32]).expect("A static array of the correct length is a known good.")
-	}
-}
-
-/// Set of test accounts.
-#[derive(RuntimeDebug, Clone, Copy)]
-pub enum TestKeyring {
-	Alice,
-	Bob,
-	Charlie,
-	Dave,
-	Eve,
-	Ferdie,
-}
-
-impl Keyring for TestKeyring {
-	fn secret(&self) -> SecretKey {
-		SecretKey::from_bytes(&[*self as u8; 32]).expect("A static array of the correct length is a known good.")
-	}
-}
-
-impl From<TestKeyring> for AuthorityId {
-	fn from(k: TestKeyring) -> Self {
-		AuthorityId::from_slice(&k.public().to_bytes())
-	}
-}
-
-impl From<Account> for AuthorityId {
-	fn from(p: Account) -> Self {
-		AuthorityId::from_slice(&p.public().to_bytes())
-	}
-}
-
-/// Get a valid set of voters for a Grandpa round.
-pub fn voter_set() -> VoterSet<AuthorityId> {
-	VoterSet::new(authority_list()).unwrap()
-}
-
-/// Convenience function to get a list of Grandpa authorities.
-pub fn authority_list() -> AuthorityList {
-	keyring().iter().map(|(id, w)| (AuthorityId::from(*id), *w)).collect()
-}
-
-/// Get the corresponding identities from the keyring for the "standard" authority set.
-pub fn keyring() -> Vec<(TestKeyring, u64)> {
-	vec![
-		(TestKeyring::Alice, 1),
-		(TestKeyring::Bob, 1),
-		(TestKeyring::Charlie, 1),
-	]
 }
